refactored Users into a separate repository
