+++ app/src/main/java/com/github/se/wanderpals/model/repository/TripsRepository.kt
@@ -254,32 +254,7 @@ open class TripsRepository(
      * @param userId The unique identifier of the user.
      * @return A `User` object if found, `null` otherwise.
      */
-  open suspend fun getUserFromTrip(tripId: String, userId: String): User? =
-      withContext(dispatcher) {
-        try {
-          val documentSnapshot =
-              tripsCollection
-                  .document(tripId)
-                  .collection(FirebaseCollections.USERS_SUBCOLLECTION.path)
-                  .document(userId)
-                  .get(getSource())
-                  .await()
-
-          val firestoreUser = documentSnapshot.toObject<FirestoreUser>()
-          if (firestoreUser != null) {
-            firestoreUser.toUser()
-          } else {
-            Log.e("TripsRepository", "getUserFromTrip: Not found user $userId from trip $tripId.")
-            null
-          }
-        } catch (e: Exception) {
-          Log.e(
-              "TripsRepository",
-              "getUserFromTrip: Error getting an user $userId from trip $tripId.",
-              e)
-          null // error
-        }
-      }
+  open suspend fun getUserFromTrip(tripId: String, userId: String): User? = userRepository.getUserFromTrip(tripId = tripId, userId = userId, source = getSource())
 
   /**
    * Retrieves all users associated with a specific trip. It iterates over all user IDs stored
@@ -290,22 +265,7 @@ open class TripsRepository(
    * @return A list of `User` objects. Returns an empty list if the trip is not found or in case of
    *   an error.
    */
-  open suspend fun getAllUsersFromTrip(tripId: String): List<User> =
-      withContext(dispatcher) {
-        try {
-          val trip = getTrip(tripId)
-          if (trip != null) {
-            val stopIds = trip.users
-            stopIds.mapNotNull { userId -> getUserFromTrip(tripId, userId) }
-          } else {
-            Log.e("TripsRepository", "getAllUsersFromTrip: Trip not found with ID $tripId.")
-            emptyList()
-          }
-        } catch (e: Exception) {
-          Log.e("TripsRepository", "getAllUsersFromTrip: Error fetching trip to trip $tripId.", e)
-          emptyList()
-        }
-      }
+  open suspend fun getAllUsersFromTrip(tripId: String): List<User> =userRepository.getAllUsersFromTrip(tripId = tripId, source = getSource())
 
   /**
    * Adds a user to a specified trip. This method creates or updates a document in the Users
@@ -315,47 +275,9 @@ open class TripsRepository(
    * @param user The `User` object containing the user's details.
    * @return `true` if the operation is successful, `false` otherwise.
    */
-  open suspend fun addUserToTrip(tripId: String, user: User): Boolean =
-      withContext(dispatcher) {
-        try {
-          if (!checkNetworkIsValidAndLog()) {
-            return@withContext false
-          }
-          // for users, there IDs are google ids currently no need to gen a new one
-          val firestoreUser = FirestoreUser.fromUser(user.copy(userId = uid))
-          val userDocument =
-              tripsCollection
-                  .document(tripId)
-                  .collection(FirebaseCollections.USERS_SUBCOLLECTION.path)
-                  .document(firestoreUser.userId)
-          userDocument.set(firestoreUser).await()
-          Log.d("TripsRepository", "addUserToTrip: User added successfully to trip $tripId.")
-          val trip = getTrip(tripId)
-          if (trip != null) {
-            // Add the new userID to the trip's user list and update the trip
-            val updatedStopsList = trip.users + user.userId
-
-            val updatedTokensList = trip.tokenIds.toMutableList()
-            if (!SessionManager.getNotificationToken().isEmpty()) {
-              updatedTokensList += listOf(SessionManager.getNotificationToken())
-              // problem when fetching the tokenIds
-              // updatedTokensList = updatedStopsList + SessionManager.getNotificationToken()
-
-            }
-            val updatedTrip = trip.copy(users = updatedStopsList, tokenIds = updatedTokensList)
-            updateTrip(updatedTrip)
-            Log.d("TripsRepository", "addUserToTrip: Stop ID added to trip successfully.")
-            true
-          } else {
-
-            Log.e("TripsRepository", "addUserToTrip: Trip not found with ID $tripId.")
-            false
-          }
-        } catch (e: Exception) {
-          Log.e("TripsRepository", "addUserToTrip: Error adding user to trip $tripId.", e)
-          false
-        }
-      }
+  open suspend fun addUserToTrip(tripId: String, user: User): Boolean =if (checkNetworkIsValidAndLog())
+      userRepository.addUserToTrip(tripId = tripId, user = user)
+  else false
 
   /**
    * Updates details of a user in a specified trip. Similar to `addUserToTrip`, but specifically
@@ -365,32 +287,9 @@ open class TripsRepository(
    * @param user The `User` object to be updated.
    * @return `true` if the operation is successful, `false` otherwise.
    */
-  open suspend fun updateUserInTrip(tripId: String, user: User): Boolean =
-      withContext(dispatcher) {
-        try {
-          Log.d("TripsRepository", "updateUserInTrip: Updating a user in trip $tripId")
-          if (!checkNetworkIsValidAndLog()) {
-            return@withContext false
-          }
-          val firestoreUser = FirestoreUser.fromUser(user)
-          tripsCollection
-              .document(tripId)
-              .collection(FirebaseCollections.USERS_SUBCOLLECTION.path)
-              .document(user.userId)
-              .set(firestoreUser)
-              .await()
-          Log.d(
-              "TripsRepository",
-              "updateUserInTrip: Trip's user updated successfully for ID $tripId.")
-          true
-        } catch (e: Exception) {
-          Log.e(
-              "TripsRepository",
-              "updateUserInTrip: Error updating user with ID ${user.userId} in trip with ID $tripId",
-              e)
-          false
-        }
-      }
+  open suspend fun updateUserInTrip(tripId: String, user: User): Boolean =if (checkNetworkIsValidAndLog())
+      userRepository.updateUserInTrip(tripId = tripId, user = user)
+  else false
 
   /**
    * Removes a user from a trip. This method deletes a user's document from the Users subcollection
@@ -400,47 +299,9 @@ open class TripsRepository(
    * @param userId The unique identifier of the user to be removed.
    * @return `true` if the operation is successful, `false` otherwise.
    */
-  open suspend fun removeUserFromTrip(tripId: String, userId: String): Boolean =
-      withContext(dispatcher) {
-        try {
-          Log.d("TripsRepository", "removeUserFromTrip: Deleting user $userId from trip $tripId")
-          if (!checkNetworkIsValidAndLog()) {
-            return@withContext false
-          }
-          tripsCollection
-              .document(tripId)
-              .collection(FirebaseCollections.USERS_SUBCOLLECTION.path)
-              .document(userId)
-              .delete()
-              .await()
-
-          val trip = getTrip(tripId)
-          if (trip != null) {
-            val updatedUsersList = trip.users.filterNot { it == userId }
-            var updatedTokensList = trip.tokenIds
-            if (trip.tokenIds.contains(SessionManager.getNotificationToken())) {
-              updatedTokensList = updatedTokensList - SessionManager.getNotificationToken()
-            }
-            val updatedTrip = trip.copy(users = updatedUsersList, tokenIds = updatedTokensList)
-
-            updateTrip(updatedTrip)
-            removeTripId(tripId, userId) // remove the Trip from the the deleted user
-            Log.d(
-                "TripsRepository",
-                "removeUserFromTrip: User $userId deleted and trip updated successfully.")
-            true
-          } else {
-            Log.e("TripsRepository", "removeUserFromTrip: Trip not found with ID $tripId.")
-            false
-          }
-        } catch (e: Exception) {
-          Log.e(
-              "TripsRepository",
-              "removeUserFromTrip: Error deleting user $userId from trip $tripId.",
-              e)
-          false
-        }
-      }
+  open suspend fun removeUserFromTrip(tripId: String, userId: String): Boolean =if (checkNetworkIsValidAndLog())
+      userRepository.removeUserFromTrip(tripId = tripId,userId = userId)
+  else false
 
   open suspend fun getStopFromTrip(tripId: String, stopId: String): Stop? = stopRepository.getStopFromTrip(tripId = tripId, stopId = stopId, source = getSource())
