+++ app/src/main/java/com/github/se/wanderpals/model/repository/UserRepository.kt
@@ -1,9 +1,13 @@
 package com.github.se.wanderpals.model.repository
 
 import android.util.Log
+import com.github.se.wanderpals.model.data.User
+import com.github.se.wanderpals.model.firestoreData.FirestoreUser
+import com.github.se.wanderpals.service.SessionManager
 import com.google.firebase.firestore.CollectionReference
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.Source
+import com.google.firebase.firestore.toObject
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.tasks.await
 import kotlinx.coroutines.withContext
@@ -116,6 +120,195 @@ open class UserRepository(
                 Log.e(
                     "TripsRepository",
                     "deleteEmailByUsername: Error deleting email for username $username.",
+                    e)
+                false
+            }
+        }
+
+
+    /**
+     * Fetches a user's details for a specific trip. This method queries a subcollection within a trip
+     * document to retrieve a user object based on the provided `userId`.
+     *
+     * @param tripId The unique identifier of the trip.
+     * @param userId The unique identifier of the user.
+     * @return A `User` object if found, `null` otherwise.
+     */
+    open suspend fun getUserFromTrip(tripId: String, userId: String,source: Source): User? =
+        withContext(dispatcher) {
+            try {
+                val documentSnapshot =
+                    tripsCollection
+                        .document(tripId)
+                        .collection(FirebaseCollections.USERS_SUBCOLLECTION.path)
+                        .document(userId)
+                        .get(source)
+                        .await()
+
+                val firestoreUser = documentSnapshot.toObject<FirestoreUser>()
+                if (firestoreUser != null) {
+                    firestoreUser.toUser()
+                } else {
+                    Log.e("TripsRepository", "getUserFromTrip: Not found user $userId from trip $tripId.")
+                    null
+                }
+            } catch (e: Exception) {
+                Log.e(
+                    "TripsRepository",
+                    "getUserFromTrip: Error getting an user $userId from trip $tripId.",
+                    e)
+                null // error
+            }
+        }
+
+    /**
+     * Retrieves all users associated with a specific trip. It iterates over all user IDs stored
+     * within a trip document and fetches their corresponding user objects.
+     *
+     * @param tripId The unique identifier of the trip. appears unused and may be a mistake in the
+     *   method signature.)
+     * @return A list of `User` objects. Returns an empty list if the trip is not found or in case of
+     *   an error.
+     */
+    open suspend fun getAllUsersFromTrip(tripId: String,source: Source): List<User> =
+        withContext(dispatcher) {
+            try {
+                val trip = tripsRepository.getTrip(tripId)
+                if (trip != null) {
+                    val stopIds = trip.users
+                    stopIds.mapNotNull { userId -> getUserFromTrip(tripId, userId,source) }
+                } else {
+                    Log.e("TripsRepository", "getAllUsersFromTrip: Trip not found with ID $tripId.")
+                    emptyList()
+                }
+            } catch (e: Exception) {
+                Log.e("TripsRepository", "getAllUsersFromTrip: Error fetching trip to trip $tripId.", e)
+                emptyList()
+            }
+        }
+
+    /**
+     * Adds a user to a specified trip. This method creates or updates a document in the Users
+     * subcollection within a trip, storing the user's details.
+     *
+     * @param tripId The unique identifier of the trip.
+     * @param user The `User` object containing the user's details.
+     * @return `true` if the operation is successful, `false` otherwise.
+     */
+    open suspend fun addUserToTrip(tripId: String, user: User): Boolean =
+        withContext(dispatcher) {
+            try {
+                // for users, there IDs are google ids currently no need to gen a new one
+                val firestoreUser = FirestoreUser.fromUser(user.copy(userId = uid))
+                val userDocument =
+                    tripsCollection
+                        .document(tripId)
+                        .collection(FirebaseCollections.USERS_SUBCOLLECTION.path)
+                        .document(firestoreUser.userId)
+                userDocument.set(firestoreUser).await()
+                Log.d("TripsRepository", "addUserToTrip: User added successfully to trip $tripId.")
+                val trip = tripsRepository.getTrip(tripId)
+                if (trip != null) {
+                    // Add the new userID to the trip's user list and update the trip
+                    val updatedStopsList = trip.users + user.userId
+
+                    val updatedTokensList = trip.tokenIds.toMutableList()
+                    if (!SessionManager.getNotificationToken().isEmpty()) {
+                        updatedTokensList += listOf(SessionManager.getNotificationToken())
+                        // problem when fetching the tokenIds
+                        // updatedTokensList = updatedStopsList + SessionManager.getNotificationToken()
+
+                    }
+                    val updatedTrip = trip.copy(users = updatedStopsList, tokenIds = updatedTokensList)
+                    tripsRepository.updateTrip(updatedTrip)
+                    Log.d("TripsRepository", "addUserToTrip: Stop ID added to trip successfully.")
+                    true
+                } else {
+
+                    Log.e("TripsRepository", "addUserToTrip: Trip not found with ID $tripId.")
+                    false
+                }
+            } catch (e: Exception) {
+                Log.e("TripsRepository", "addUserToTrip: Error adding user to trip $tripId.", e)
+                false
+            }
+        }
+
+    /**
+     * Updates details of a user in a specified trip. Similar to `addUserToTrip`, but specifically
+     * intended for updating existing user documents.
+     *
+     * @param tripId The unique identifier of the trip.
+     * @param user The `User` object to be updated.
+     * @return `true` if the operation is successful, `false` otherwise.
+     */
+    open suspend fun updateUserInTrip(tripId: String, user: User): Boolean =
+        withContext(dispatcher) {
+            try {
+                Log.d("TripsRepository", "updateUserInTrip: Updating a user in trip $tripId")
+
+                val firestoreUser = FirestoreUser.fromUser(user)
+                tripsCollection
+                    .document(tripId)
+                    .collection(FirebaseCollections.USERS_SUBCOLLECTION.path)
+                    .document(user.userId)
+                    .set(firestoreUser)
+                    .await()
+                Log.d(
+                    "TripsRepository",
+                    "updateUserInTrip: Trip's user updated successfully for ID $tripId.")
+                true
+            } catch (e: Exception) {
+                Log.e(
+                    "TripsRepository",
+                    "updateUserInTrip: Error updating user with ID ${user.userId} in trip with ID $tripId",
+                    e)
+                false
+            }
+        }
+
+    /**
+     * Removes a user from a trip. This method deletes a user's document from the Users subcollection
+     * within a trip. And the TripId from his accessible trips
+     *
+     * @param tripId The unique identifier of the trip.
+     * @param userId The unique identifier of the user to be removed.
+     * @return `true` if the operation is successful, `false` otherwise.
+     */
+    open suspend fun removeUserFromTrip(tripId: String, userId: String): Boolean =
+        withContext(dispatcher) {
+            try {
+                Log.d("TripsRepository", "removeUserFromTrip: Deleting user $userId from trip $tripId")
+                tripsCollection
+                    .document(tripId)
+                    .collection(FirebaseCollections.USERS_SUBCOLLECTION.path)
+                    .document(userId)
+                    .delete()
+                    .await()
+
+                val trip = tripsRepository.getTrip(tripId)
+                if (trip != null) {
+                    val updatedUsersList = trip.users.filterNot { it == userId }
+                    var updatedTokensList = trip.tokenIds
+                    if (trip.tokenIds.contains(SessionManager.getNotificationToken())) {
+                        updatedTokensList = updatedTokensList - SessionManager.getNotificationToken()
+                    }
+                    val updatedTrip = trip.copy(users = updatedUsersList, tokenIds = updatedTokensList)
+
+                    tripsRepository.updateTrip(updatedTrip)
+                    tripsRepository.removeTripId(tripId, userId) // remove the Trip from the the deleted user
+                    Log.d(
+                        "TripsRepository",
+                        "removeUserFromTrip: User $userId deleted and trip updated successfully.")
+                    true
+                } else {
+                    Log.e("TripsRepository", "removeUserFromTrip: Trip not found with ID $tripId.")
+                    false
+                }
+            } catch (e: Exception) {
+                Log.e(
+                    "TripsRepository",
+                    "removeUserFromTrip: Error deleting user $userId from trip $tripId.",
                     e)
                 false
             }
